# Faster random() scalar function