--- a/datafusion/functions/Cargo.toml
+++ b/datafusion/functions/Cargo.toml
@@ -156,3 +156,8 @@ required-features = ["unicode_expressions"]
 harness = false
 name = "repeat"
 required-features = ["string_expressions"]
+
+[[bench]]
+harness = false
+name = "random"
+required-features = ["math_expressions"]